requirements.txt
asn1crypto==1.5.1
certifi==2023.7.22
certvalidator==0.11.1
cffi==1.15.1
charset-normalizer==3.2.0
idna==3.4
lief==0.12.3
mscerts==2023.8.23
oscrypto==1.3.0
pefile==2023.2.7
peid==1.2.9
pyasn1==0.4.8
pyasn1-modules==0.3.0
pycparser==2.21
python-magic==0.4.27
requests==2.31.0
signify==0.5.2
six==1.16.0
ssdeep==3.4
urllib3==2.0.5
xmltodict==0.13.0
yara-python==4.3.1
